# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Smart Biometric Attendance System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  studentId: string;
  name: string;
  className: string;
  enrolledAt: string;
}

export interface StudentInput {
  /** @minLength 1 */
  studentId: string;
  /** @minLength 1 */
  name: string;
  className?: string;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  'on-time': 'on-time',
  late: 'late',
} as const;

export interface AttendanceRecord {
  id: number;
  studentId: number;
  studentName: string;
  className: string;
  date: string;
  timestamp: string;
  status: AttendanceRecordStatus;
}

export interface ScanInput {
  studentId: number;
}

export interface AlreadyScannedResponse {
  message: string;
  record: AttendanceRecord;
}

export interface AttendanceStats {
  date: string;
  totalStudents: number;
  present: number;
  absent: number;
  onTime: number;
  late: number;
}

export interface DailyCount {
  date: string;
  present: number;
  late: number;
  onTime: number;
}

export interface ClassAttendance {
  className: string;
  present: number;
  absent: number;
  total: number;
}

/**
 * Raw WebAuthn options object (PublicKeyCredentialCreationOptionsJSON or RequestOptionsJSON)
 */
export interface WebAuthnOptionsPayload {
  challenge: string;
  [key: string]: unknown;
 }

export interface FaceIdStudentRef {
  studentId: number;
}

/**
 * RegistrationResponseJSON from @simplewebauthn/browser
 */
export type FaceIdEnrollmentCredential = {
  id: string;
  rawId: string;
  type: string;
  [key: string]: unknown;
 };

export interface FaceIdEnrollment {
  studentId: number;
  /** RegistrationResponseJSON from @simplewebauthn/browser */
  credential: FaceIdEnrollmentCredential;
}

/**
 * AuthenticationResponseJSON from @simplewebauthn/browser
 */
export type FaceIdVerificationCredential = {
  id: string;
  rawId: string;
  type: string;
  [key: string]: unknown;
 };

export interface FaceIdVerification {
  /** AuthenticationResponseJSON from @simplewebauthn/browser */
  credential: FaceIdVerificationCredential;
}

export interface FaceIdCredential {
  id: number;
  studentId: number;
  createdAt: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type ListAttendanceParams = {
date?: string;
};

export type GetAttendanceStatsParams = {
date?: string;
};

export type GetAttendanceByClassParams = {
date?: string;
};