# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Biometric Attendance System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: students
    description: Student enrollment operations
  - name: attendance
    description: Attendance scanning and logging operations
  - name: webauthn
    description: Face ID / biometric credential operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all enrolled students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: enrollStudent
      tags: [students]
      summary: Enroll a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Student ID already enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a single student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Remove a student from enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records for a date
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"

  /attendance/scan:
    post:
      operationId: simulateScan
      tags: [attendance]
      summary: Simulate a fingerprint scan for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Scan logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "200":
          description: Student already scanned today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlreadyScannedResponse"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /attendance/stats:
    get:
      operationId: getAttendanceStats
      tags: [attendance]
      summary: Get daily attendance statistics
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Stats for the day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStats"

  /attendance/weekly:
    get:
      operationId: getWeeklyAttendance
      tags: [attendance]
      summary: Get attendance counts for the last 7 days
      responses:
        "200":
          description: Weekly attendance breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyCount"

  /attendance/by-class:
    get:
      operationId: getAttendanceByClass
      tags: [attendance]
      summary: Get present/absent counts grouped by class for today
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Per-class attendance breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassAttendance"

  /webauthn/register-options:
    post:
      operationId: getFaceIdRegisterOptions
      tags: [webauthn]
      summary: Get WebAuthn registration options for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaceIdStudentRef"
      responses:
        "200":
          description: Registration options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebAuthnOptionsPayload"
        "404":
          description: Student not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /webauthn/register:
    post:
      operationId: completeFaceIdRegister
      tags: [webauthn]
      summary: Complete Face ID registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaceIdEnrollment"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /webauthn/auth-options:
    get:
      operationId: getFaceIdAuthOptions
      tags: [webauthn]
      summary: Get WebAuthn authentication options
      responses:
        "200":
          description: Authentication options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebAuthnOptionsPayload"

  /webauthn/auth:
    post:
      operationId: completeFaceIdAuth
      tags: [webauthn]
      summary: Complete Face ID authentication and record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaceIdVerification"
      responses:
        "201":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "200":
          description: Already scanned today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlreadyScannedResponse"
        "400":
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Device not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /webauthn/credentials/{studentId}:
    get:
      operationId: listFaceIdCredentials
      tags: [webauthn]
      summary: List registered Face ID credentials for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Credentials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaceIdCredential"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Student:
      type: object
      required: [id, studentId, name, className, enrolledAt]
      properties:
        id:
          type: integer
        studentId:
          type: string
        name:
          type: string
        className:
          type: string
        enrolledAt:
          type: string

    StudentInput:
      type: object
      required: [studentId, name]
      properties:
        studentId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        className:
          type: string

    AttendanceRecord:
      type: object
      required: [id, studentId, studentName, className, date, timestamp, status]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        className:
          type: string
        date:
          type: string
        timestamp:
          type: string
        status:
          type: string
          enum: [on-time, late]

    ScanInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer

    AlreadyScannedResponse:
      type: object
      required: [message, record]
      properties:
        message:
          type: string
        record:
          $ref: "#/components/schemas/AttendanceRecord"

    AttendanceStats:
      type: object
      required: [date, totalStudents, present, absent, onTime, late]
      properties:
        date:
          type: string
        totalStudents:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        onTime:
          type: integer
        late:
          type: integer

    DailyCount:
      type: object
      required: [date, present, late, onTime]
      properties:
        date:
          type: string
        present:
          type: integer
        late:
          type: integer
        onTime:
          type: integer

    ClassAttendance:
      type: object
      required: [className, present, absent, total]
      properties:
        className:
          type: string
        present:
          type: integer
        absent:
          type: integer
        total:
          type: integer

    # WebAuthn / Face ID schemas
    WebAuthnOptionsPayload:
      type: object
      description: Raw WebAuthn options object (PublicKeyCredentialCreationOptionsJSON or RequestOptionsJSON)
      properties:
        challenge:
          type: string
      additionalProperties: true
      required: [challenge]

    FaceIdStudentRef:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer

    FaceIdEnrollment:
      type: object
      required: [studentId, credential]
      properties:
        studentId:
          type: integer
        credential:
          type: object
          description: RegistrationResponseJSON from @simplewebauthn/browser
          properties:
            id:
              type: string
            rawId:
              type: string
            type:
              type: string
          additionalProperties: true
          required: [id, rawId, type]

    FaceIdVerification:
      type: object
      required: [credential]
      properties:
        credential:
          type: object
          description: AuthenticationResponseJSON from @simplewebauthn/browser
          properties:
            id:
              type: string
            rawId:
              type: string
            type:
              type: string
          additionalProperties: true
          required: [id, rawId, type]

    FaceIdCredential:
      type: object
      required: [id, studentId, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        createdAt:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string