# lib/api-zod/src/generated/types/attendanceStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Smart Biometric Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceStats {
  date: string;
  totalStudents: number;
  present: number;
  absent: number;
  onTime: number;
  late: number;
}