# lib/api-zod/src/generated/types/faceIdStudentRef.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Smart Biometric Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface FaceIdStudentRef {
  studentId: number;
}